Finished Crouching
I lied. There's still setting the crouch speed for how fast you are moving when crouched but that's it. Everything else is done. Which is kidn of cool because I have never crouched before. I'm hoping this is as dam mdoular as I can make it. Tried to avoid absolutely every hardcode possible.