Create()
Got the create method to work. Gonna get everything flowing before making enhancements. For example, I am going to remove the hard coded transforms and have all the stacks and lists auto generate as needed. Then, I'm going to auto "pre-initialize' them at start for all the enums. Also going to move the enum logic away from Create and have it be in the Poolable script. Poolable is going to be a helper script for the pool. It'll allow each object to keep track of itself from the perspective of the pool.