Spawn Manager
Finally had time to actually sit down and think about the SpawnManaegr. The previous version was mostly just  to get the test in. NOW it's for real for real becoming modular with some thought. Creating several methods to allow the WaveManager to have many options of What, QTY, Where, and interval.